refactor!: Move constraint circuits to own crate
changelog: ignore
