+++ triton-vm/Cargo.toml
@@ -40,6 +40,7 @@ syn.workspace = true
 thiserror.workspace = true
 twenty-first.workspace = true
 unicode-width.workspace = true
+constraint-builder = { path = "../constraint-builder" }
 
 [dev-dependencies]
 assert2.workspace = true
