+++ triton-vm/src/table/cross_table_argument.rs
@@ -1,11 +1,11 @@
 use std::ops::Add;
 use std::ops::Mul;
 
-use crate::codegen::circuit::ConstraintCircuitBuilder;
-use crate::codegen::circuit::ConstraintCircuitMonad;
-use crate::codegen::circuit::DualRowIndicator;
-use crate::codegen::circuit::SingleRowIndicator;
-use crate::codegen::circuit::SingleRowIndicator::ExtRow;
+use constraint_builder::ConstraintCircuitBuilder;
+use constraint_builder::ConstraintCircuitMonad;
+use constraint_builder::DualRowIndicator;
+use constraint_builder::SingleRowIndicator;
+use constraint_builder::SingleRowIndicator::ExtRow;
 use twenty_first::prelude::*;
 
 use crate::table::challenges::ChallengeId::*;
