+++ triton-vm/src/table/hash_table.rs
@@ -1,10 +1,10 @@
-use crate::codegen::circuit::ConstraintCircuitBuilder;
-use crate::codegen::circuit::ConstraintCircuitMonad;
-use crate::codegen::circuit::DualRowIndicator;
-use crate::codegen::circuit::DualRowIndicator::*;
-use crate::codegen::circuit::InputIndicator;
-use crate::codegen::circuit::SingleRowIndicator;
-use crate::codegen::circuit::SingleRowIndicator::*;
+use constraint_builder::ConstraintCircuitBuilder;
+use constraint_builder::ConstraintCircuitMonad;
+use constraint_builder::DualRowIndicator;
+use constraint_builder::DualRowIndicator::*;
+use constraint_builder::InputIndicator;
+use constraint_builder::SingleRowIndicator;
+use constraint_builder::SingleRowIndicator::*;
 use itertools::Itertools;
 use ndarray::*;
 use num_traits::Zero;
