+++ triton-vm/src/table/processor_table.rs
@@ -1,9 +1,9 @@
 use std::cmp::max;
 use std::ops::Mul;
 
-use crate::codegen::circuit::DualRowIndicator::*;
-use crate::codegen::circuit::SingleRowIndicator::*;
-use crate::codegen::circuit::*;
+use constraint_builder::DualRowIndicator::*;
+use constraint_builder::SingleRowIndicator::*;
+use constraint_builder::*;
 use itertools::izip;
 use itertools::Itertools;
 use ndarray::parallel::prelude::*;
