+++ triton-vm/src/table/program_table.rs
@@ -1,8 +1,8 @@
 use std::cmp::Ordering;
 
-use crate::codegen::circuit::DualRowIndicator::*;
-use crate::codegen::circuit::SingleRowIndicator::*;
-use crate::codegen::circuit::*;
+use constraint_builder::DualRowIndicator::*;
+use constraint_builder::SingleRowIndicator::*;
+use constraint_builder::*;
 use ndarray::s;
 use ndarray::Array1;
 use ndarray::ArrayView1;
