+++ triton-vm/src/table/ram_table.rs
@@ -1,9 +1,9 @@
 use std::cmp::Ordering;
 
-use crate::codegen::circuit::DualRowIndicator::*;
-use crate::codegen::circuit::SingleRowIndicator::*;
-use crate::codegen::circuit::*;
 use arbitrary::Arbitrary;
+use constraint_builder::DualRowIndicator::*;
+use constraint_builder::SingleRowIndicator::*;
+use constraint_builder::*;
 use itertools::Itertools;
 use ndarray::parallel::prelude::*;
 use ndarray::prelude::*;
